requirements.txt
beautifulsoup4==4.12.0
google_api_python_client==2.86.0
protobuf==4.22.3
Requests==2.29.0
youtube_transcript_api==0.6.0
